edit student, bug them nhieu lan

+++ public/js/class/Render.js
@@ -7,7 +7,7 @@ export class Render {
     removeAllRowInTable() {
         let tbody = document.getElementById("data-students");
         let arrRow = tbody.getElementsByTagName("tr");
-        for (let i = 1; i < arrRow.length;) {
+        for (let i = 0; i < arrRow.length;) {
             arrRow[i].remove();
         }
     }
@@ -44,8 +44,35 @@ export class Render {
             listStudent = [];
         }
     }
-    addStudent(listStudent, buttonAdd, add, showForm) {
-        let formAddStudent = document.getElementById("form-add-student");
+    createForm(tbody) {
+        let tr = document.createElement("tr");
+        let tdList = [];
+        for (let i = 1; i <= 9; i++) {
+            let td = document.createElement("td");
+            tdList.push(td);
+        }
+        tr.appendChild(tdList[0]);
+        tr.appendChild(tdList[1]);
+        tdList[2].innerHTML = `<input type="text" name="name">`;
+        tr.appendChild(tdList[2]);
+        tdList[3].innerHTML = `<input type="text" name="username">`;
+        tr.appendChild(tdList[3]);
+        tdList[4].innerHTML = `<input type="checkbox" name="status">`;
+        tr.appendChild(tdList[4]);
+        tdList[5].innerHTML = `<input type="radio" name="role" id="admin">
+                        <label for="admin">Admin</label>
+                        <input type="radio" name="role" id="user">
+                        <label for="user">User</label>`;
+        tr.appendChild(tdList[5]);
+        tdList[6].innerHTML = `<input type="email" name="email">`;
+        tr.appendChild(tdList[6]);
+        tdList[7].innerHTML = `<input type="date" name="date">`;
+        tr.appendChild(tdList[7]);
+        tdList[8].innerHTML = `<input type="text" name="idStudent">`;
+        tr.appendChild(tdList[8]);
+        tbody.insertBefore(tr, tbody.firstChild);
+    }
+    addStudent(listStudent) {
         let name = document.getElementsByName("name")[0].value;
         let username = document.getElementsByName("username")[0].value;
         ;
@@ -55,11 +82,8 @@ export class Render {
         let date = document.getElementsByName("date")[0].value;
         let id = document.getElementsByName("idStudent")[0].value;
         let s = new Student(id, name, username, status, role, email, new Date(date));
-        listStudent.push(s);
         try {
             this.studentService.insertStudent(s);
-            buttonAdd.removeEventListener("click", add);
-            buttonAdd.addEventListener("click", showForm);
             alert("Luu thanh cong");
         }
         catch (error) {
@@ -84,7 +108,7 @@ export class Render {
             }
         }
     }
-    deleteStudent() {
+    getAllIdStudentSelect() {
         let allInputDelete = document.getElementsByName("edit-and-delete");
         let listIdStudent = [];
         for (let i = 0; i < allInputDelete.length; i++) {
@@ -93,22 +117,72 @@ export class Render {
                 listIdStudent.push(inputHTML.getAttribute("id"));
             }
         }
-        listIdStudent.forEach(id => {
-            this.studentService.deleteStudent(id);
-        });
+        return listIdStudent;
     }
-    showForm(buttonAdd, buttonDelete, add, showForm, deleteSt, hidenForm) {
-        let formAddStudent = document.getElementById("form-add-student");
-        formAddStudent.setAttribute("class", "form-add-student-show");
+    deleteStudent() {
+        let listIdStudent = this.getAllIdStudentSelect();
+        if (listIdStudent.length > 0) {
+            listIdStudent.forEach(id => {
+                this.studentService.deleteStudent(id);
+            });
+        }
+    }
+    showStudentEdit() {
+        let listIdStudent = this.getAllIdStudentSelect();
+        if (listIdStudent.length > 0) {
+            let student = this.studentService.getStudnetById(listIdStudent[0]);
+            document.getElementsByName("name")[0].value = student.name;
+            document.getElementsByName("username")[0].value = student.username;
+            document.getElementsByName("status")[0].checked = student.status;
+            if (student.role === "Admin") {
+                document.getElementsByName("role")[0].checked = true;
+            }
+            else {
+                document.getElementsByName("role")[1].checked = true;
+            }
+            document.getElementsByName("email")[0].value = student.email;
+            document.getElementsByName("date")[0].valueAsDate = new Date(student.dateActive);
+            let inputId = document.getElementsByName("idStudent")[0];
+            inputId.setAttribute("readonly", "true");
+            inputId.value = student.id;
+        }
+    }
+    editStudent() {
+        let name = document.getElementsByName("name")[0].value;
+        let username = document.getElementsByName("username")[0].value;
+        ;
+        let status = document.getElementsByName("status")[0].checked;
+        let role = document.getElementsByName("role")[0].checked ? "Admin" : "User";
+        let email = document.getElementsByName("email")[0].value;
+        let date = document.getElementsByName("date")[0].value;
+        let id = document.getElementsByName("idStudent")[0].value;
+        let s = new Student(id, name, username, status, role, email, new Date(date));
+        try {
+            this.studentService.updateStudent(s);
+            alert("Cap nhat thanh cong");
+        }
+        catch (error) {
+            alert("Khong the cap nhat");
+        }
+    }
+    showForm(buttonAdd, add, showForm, buttonDelete, deleteSt, hidenForm, btnEdit, edit, showData) {
+        let tbody = document.getElementById("data-students");
+        this.createForm(tbody);
         buttonAdd.removeEventListener("click", showForm);
         buttonAdd.addEventListener("click", add);
         buttonDelete.removeEventListener("click", deleteSt);
         buttonDelete.addEventListener("click", hidenForm);
+        btnEdit.removeEventListener("click", showData);
+        btnEdit.addEventListener("click", edit);
     }
-    hidenForm(buttonDelete, deleteSt, hidenForm) {
-        let formAddStudent = document.getElementById("form-add-student");
-        formAddStudent.setAttribute("class", "form-add-student-hiden");
+    hidenForm(buttonAdd, add, showForm, buttonDelete, deleteSt, hidenForm, btnEdit, edit, showData) {
+        let tbody = document.getElementById("data-students");
+        tbody.removeChild(tbody.firstChild);
+        buttonAdd.removeEventListener("click", add);
+        buttonAdd.addEventListener("click", showForm);
         buttonDelete.removeEventListener("click", hidenForm);
         buttonDelete.addEventListener("click", deleteSt);
+        btnEdit.removeEventListener("click", edit);
+        btnEdit.addEventListener("click", showData);
     }
 }
